requirements.txt
pandas~=1.5.1
PyYAML~=6.0
genpipes~=0.1.2